// types/NewsType.ts
export default interface NewsType {
	id: string;
	sections: string[];
	title: string;
	publisher: string;
	author: string;
	summary: string;
	highlight: null;
	score: null;
	image_url: string;
	thumbnail_url: string;
	content_url: string;
	esg: null;
	companies: string[];
	entities: string[];
	published_at: string;
	body: null;
}